+++ src/NServiceBus.Core.Tests/Serializers/MessageDeserializerResolverTests.cs
@@ -4,7 +4,6 @@
     using NServiceBus.MessageInterfaces.MessageMapper.Reflection;
     using NServiceBus.Serialization;
     using NServiceBus.Serializers;
-    using NServiceBus.Serializers.Binary;
     using NServiceBus.Serializers.Json;
     using NServiceBus.Serializers.XML;
     using NUnit.Framework;
@@ -21,16 +20,11 @@
             var mapper = new MessageMapper();
             var xml = new XmlMessageSerializer(mapper, new Conventions());
             var json = new JsonMessageSerializer(mapper);
-            var bson = new BsonMessageSerializer(mapper);
-            var binary = new BinaryMessageSerializer();
-
-            resolver = new MessageDeserializerResolver(new IMessageSerializer[] { xml, json, bson, binary }, xml.GetType());
+            resolver = new MessageDeserializerResolver(new IMessageSerializer[] { xml, json}, xml.GetType());
         }
 
         [TestCase(ContentTypes.Xml, typeof(XmlMessageSerializer))]
         [TestCase(ContentTypes.Json, typeof(JsonMessageSerializer))]
-        [TestCase(ContentTypes.Bson, typeof(BsonMessageSerializer))]
-        [TestCase(ContentTypes.Binary, typeof(BinaryMessageSerializer))]
         public void RetrievesSerializerByContentType(string contentType, Type expected)
         {
             var serializer = resolver.Resolve(contentType);
