+++ src/NServiceBus.Core/NServiceBus.Core.csproj
@@ -395,8 +395,6 @@
     <Compile Include="Serialization\SerializationConfigExtensions.cs" />
     <Compile Include="Serialization\SerializationDefinition.cs" />
     <Compile Include="Serialization\SerializationExtentions.cs" />
-    <Compile Include="Serializers\Binary\Binary.cs" />
-    <Compile Include="Serializers\Json\BsonSerializer.cs" />
     <Compile Include="Serializers\Json\JsonSerializer.cs" />
     <Compile Include="Serializers\XML\Config\XmlSerializer.cs" />
     <Compile Include="Serializers\XML\Config\XmlSerializationExtentions.cs" />
@@ -531,13 +529,11 @@
     <Compile Include="UnitOfWork\UnitOfWorkBehavior.cs" />
     <Compile Include="Pipeline\Behavior.cs" />
     <Compile Include="Pipeline\BehaviorChain.cs" />
-    <Compile Include="Serializers\Binary\Config\BinarySerialization.cs" />
     <Compile Include="CriticalError\ConfigureCriticalErrorAction.cs" />
     <Compile Include="DataBus\ConfigureFileShareDataBus.cs" />
     <Compile Include="Features\FeatureActivator.cs" />
     <Compile Include="Recoverability\SecondLevelRetries\Config\SecondLevelRetriesConfigExtensions.cs" />
     <Compile Include="Recoverability\SecondLevelRetries\Config\SecondLevelRetriesSettings.cs" />
-    <Compile Include="Serializers\Json\Config\BsonSerialization.cs" />
     <Compile Include="Serializers\Json\Config\JsonSerialization.cs" />
     <Compile Include="Serializers\Json\Config\JsonSerializerConfigurationExtensions.cs" />
     <Compile Include="Serializers\XML\Config\XmlSerialization.cs" />
@@ -620,16 +616,10 @@
     <Compile Include="Scheduling\InMemoryScheduledTaskStorage.cs" />
     <Compile Include="Scheduling\Messages\ScheduledTask.cs" />
     <Compile Include="Scheduling\ScheduledTaskMessageHandler.cs" />
-    <Compile Include="Serializers\Binary\BinaryMessageSerializer.cs" />
-    <Compile Include="Serializers\Binary\SimpleMessageMapper.cs" />
-    <Compile Include="Serializers\Binary\XContainerSurrogate.cs" />
-    <Compile Include="Serializers\Binary\XElementSurrogate.cs" />
-    <Compile Include="Serializers\Json\BsonMessageSerializer.cs" />
     <Compile Include="Serializers\Json\Internal\MessageContractResolver.cs" />
     <Compile Include="Serializers\Json\Internal\MessageSerializationBinder.cs" />
     <Compile Include="Serializers\Json\Internal\XContainerConverter.cs" />
     <Compile Include="Serializers\Json\JsonMessageSerializer.cs" />
-    <Compile Include="Serializers\Json\JsonMessageSerializerBase.cs" />
     <Compile Include="Serializers\XML\XmlMessageSerializer.cs" />
     <Compile Include="Serializers\XML\XmlSanitizingStream.cs" />
     <Compile Include="Settings\TransactionSettings.cs" />
