--- src/NServiceBus.Core/Serializers/Binary/Binary.cs
@@ -1,20 +1,0 @@
-namespace NServiceBus
-{
-    using System;
-    using Features;
-    using Serialization;
-
-    /// <summary>
-    /// Defines the capabilities of the Binary serializer.
-    /// </summary>
-    public class BinarySerializer : SerializationDefinition
-    {
-        /// <summary>
-        /// <see cref="SerializationDefinition.ProvidedByFeature"/>.
-        /// </summary>
-        protected internal override Type ProvidedByFeature()
-        {
-            return typeof(BinarySerialization);
-        }
-    }
-}
